fix more warnings
I previously overlooked a couple unused-but-not-set-variable
warnings in the vtkNewickTreeReader because my local version
of GCC isn't new enough to detect these.  Per Dave Demarle's
request, I also fixed a couple signed/unsigned comparison
warnings in vtkBoostDividedEdgeBundling.cxx.  There still
remains an array-bounds warnings in this file, but I'm not
comfortable attempting to fix it at this time.

+++ Infovis/BoostGraphAlgorithms/vtkBoostDividedEdgeBundling.cxx
@@ -147,9 +147,9 @@ void vtkBundlingMetadata::DenormalizeNodePositions()
       p[1] / 1000.0f * this->Scale + this->YRange[0],
       p[2] / 1000.0f * this->Scale + this->ZRange[0]);
     }
-  for (vtkIdType i = 0; i < this->EdgeMesh.size(); ++i)
+  for (vtkIdType i = 0; i < (int)this->EdgeMesh.size(); ++i)
     {
-    for (vtkIdType j = 0; j < this->EdgeMesh[i].size(); ++j)
+    for (vtkIdType j = 0; j < (int)this->EdgeMesh[i].size(); ++j)
       {
       vtkVector3f p = this->EdgeMesh[i][j];
       this->EdgeMesh[i][j] = vtkVector3f(
